custom messsages

+++ api/src/routes/book.routes.ts
@@ -49,7 +49,7 @@ app.post("/favorites", async (req, res) => {
     });
     return res.status(201).send();
   }
-  
+
   throw new AppError("Livro já favoritado!");
 });
 
@@ -58,6 +58,15 @@ app.delete("/favorites/:id", async (req, res) => {
 
   const { id } = req.params;
 
+  const verifyFavoriteIdBook = await prisma.favorites.findUnique({
+    where: {
+      id
+    }
+  })
+  if (!verifyFavoriteIdBook) {
+    return res.status(404).send({ message: "Livro não favoritado!" })
+
+  }
   await prisma.favorites.delete({
     where: {
       id,
